// sdk/src/types/aws-cli.types.ts
export interface GlueJob {
  name: string;
  description: string;
  role: string;
  createdOn: string;
  lastModifiedOn: string;
  executionProperty: unknown;
}
// ...
export interface ECRRepository {
  repositoryName: string;
  repositoryArn: string;
  registryId: string;
  createdAt: string;
}